Move old llvm2graph implementation into legacy.
This moves the old llvm2graph module into a 'legacy' pacakge in the
new implementation. I am not ready to remove it yet since there are a
few old modules that need to be updated to the new version, but this
adds an explicit depracation warning when calling the old graph
builder.

Signed-off-by: format 2020.01.16 <github.com/ChrisCummins/format>