labm8: Make labm8 work without workspace_status.
The workspace_status file is generated by bazel using a genrule(), and
contains transient info such as timestamps and user info. When
deploying as a pip package, the workspace_status file is not exported
because it is not recognized as a python library.

One solution would be to correctly emulate the py_library interface
from workspace_status. Another, quick workaround is to simply catch the
import error. This patch does the latter.

Signed-off-by: format 20.02.05 <github.com/ChrisCummins/format>